pyproject.toml
[project]
name = "obj-reid-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "open-clip-torch>=2.24.0",
    "timm>=0.9.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
]
